♻️ Extract Interface for Model
Need an interface, so that downstream projects may add their own models, instead of being restricted to the standard models included in codemodder-java.
