+++ plugins/codemodder-plugin-llm/src/main/java/io/codemodder/plugins/llm/StandardModel.java
@@ -1,0 +1,50 @@
+package io.codemodder.plugins.llm;
+
+import com.knuddels.jtokkit.api.EncodingType;
+import com.theokanning.openai.completion.chat.ChatMessage;
+import java.util.List;
+
+/** Well-known GPT models used in Codemod development. */
+public enum StandardModel implements Model {
+  GPT_3_5_TURBO("gpt-3.5-turbo-0125", 16_385) {
+    @Override
+    public int tokens(final List<ChatMessage> messages) {
+      return Tokens.countTokens(messages, 3, EncodingType.CL100K_BASE);
+    }
+  },
+  GPT_4("gpt-4-0613", 8_192) {
+    @Override
+    public int tokens(final List<ChatMessage> messages) {
+      return Tokens.countTokens(messages, 3, EncodingType.CL100K_BASE);
+    }
+  },
+  GPT_4O("gpt-4o-2024-05-13", 128_000) {
+    /**
+     * Will add support for GPT-4o token counting based on <a
+     * href="https://github.com/knuddelsgmbh/jtokkit/issues/96">pending work from upstream
+     * utility</a>.
+     */
+    @Override
+    public int tokens(final List<ChatMessage> messages) {
+      throw new UnsupportedOperationException("Cannot yet estimate tokens for GPT-4o");
+    }
+  };
+
+  private final String id;
+  private final int contextWindow;
+
+  StandardModel(final String id, final int contextWindow) {
+    this.id = id;
+    this.contextWindow = contextWindow;
+  }
+
+  @Override
+  public String id() {
+    return id;
+  }
+
+  @Override
+  public int contextWindow() {
+    return contextWindow;
+  }
+}
